+++ private/buf/cmd/buf/command/registry/module/modulesettings/modulesettingsupdate/usage.gen.go
@@ -14,6 +14,6 @@
 
 // Generated. DO NOT EDIT.
 
-package modulesettingupdate
+package modulesettingsupdate
 
 import _ "github.com/bufbuild/buf/private/usage"